Stream/server/requirements.txt
hcsr04sensor
flask
flask_cors
gunicorn
adafruit-circuitpython-dht
mysql-connector-python
RPi.GPIO
ADS1x15-ADC
